+++ src/db/word.ts
@@ -3,6 +3,7 @@ import useUserProfileStore from "@stores/userProfile";
 import { doc, getDoc, setDoc, getFirestore } from "firebase/firestore";
 import useSWR from "swr";
 import "@auth";
+import { getUserId } from "@auth";
 
 const firestore = getFirestore();
 
@@ -16,8 +17,13 @@ export async function fetchWord(request: WordRequest): Promise<Word[]> {
     return [];
   }
 
-  const words: Word[] = [];
-  for (const wordStr of request.words) {
+  const words: Word[] = new Array<Word>(request.words.length).fill(
+    Word.New(""),
+  );
+  const allPromises: Promise<void>[] = [];
+
+  for (let i = 0; i < request.words.length; i++) {
+    const wordStr = request.words[i];
     const docRef = doc(
       firestore,
       "users",
@@ -25,22 +31,45 @@ export async function fetchWord(request: WordRequest): Promise<Word[]> {
       "words",
       Word.hash(wordStr),
     );
-    const docSnap = await getDoc(docRef);
-
-    if (docSnap.exists()) {
-      const word = docSnap.data() as Word;
-      words.push(word);
-    } else {
-      words.push(Word.New(wordStr));
-    }
+    allPromises.push(
+      getDoc(docRef)
+        .then((docSnap) => {
+          if (docSnap.exists()) {
+            const word = docSnap.data() as any;
+            // Convert firebase timestamps to Date objects
+            words[i] = {
+              ...word,
+              highestWpmDatetime: word.highestWpmDatetime.toDate(),
+              lastPracticeDatetime: word.lastPracticeDatetime.toDate(),
+            } as Word;
+          } else {
+            words[i] = Word.New(wordStr);
+          }
+        })
+        .catch((error) => {
+          console.error("Error getting document:", error);
+          words[i] = Word.New(wordStr);
+        }),
+    );
   }
 
+  await Promise.all(allPromises);
   return words;
 }
 
 export async function saveWord(uid: string, word: Word): Promise<void> {
   const docRef = doc(firestore, "users", uid, "words", Word.hash(word.word));
   return setDoc(docRef, word);
+}
+
+export async function getWords(words: string[]): Promise<Word[]> {
+  const uid = getUserId();
+  if (!uid) {
+    return [];
+  }
+
+  const wordRequest: WordRequest = { uid, words };
+  return fetchWord(wordRequest);
 }
 
 export function useWords(words: string[]) {